moving external command to separate class
